[engine] Fix a case of wrong status after a retry
When the task has a first attempt with 10 tries, and then a manual retry
without any automatic retry, doing a second manual retry would put the
task in a OK status because it lost the fact that the first attempt
encountered a warning situation.

As soon as a retry happen, the task will now be considered in a warning
status to avoid edge cases like this.