Add table prefix to storage classes
This commit introduces a table prefix to all major storage classes (PostgresKeyValueStorage, PostgresKeySetStorage, MySQLKeyValueStorage and MySQLKeySetStorage). This allows for more customized naming of tables, which can reduce the risk of naming clashes in the database. It also provides an additional layer of organization for managing database tables. This change includes updates to existing SQL queries to incorporate the new table prefix.