Caching is now part of Storage (#247)
* Update storage configurations with builder methods

This update modifies the storage configuration files (Redis, InMemory, MySQL, Postgres) and provides builder methods to simplify object creation. This change is beneficial for Java users.

* Update cache configurations with builder methods

This update modifies the cache configuration files (Caffeine) and provides builder methods to simplify object creation. This change is beneficial for Java users.

* Caching is now part of Storage.

To be more consistent, the cache configuration is now part of the storage configuration